Remove duplicate code for prefered language retrieval
