+++ src/routes/statics.js
@@ -2,24 +2,15 @@ var express = require('express');
 var router  = express.Router();
 var db = require('../database');
 var sqlHelper = require('../lib/sql-helper');
+var langHelper = require('../lib/lang-helper');
 
 router.VEHICLE_TYPE = 1;
 router.FOOD_TYPE = 2;
 router.FACILITY_TYPE = 3;
 router.FUEL_TYPE = 4;
 
-var getRequestLanguage = function(req) {
-	var acceptLanguage = require('accept-language');
-	acceptLanguage.languages(['en-US', 'th-TH']);
-	var language = req.get('Accept-Language');
-	language = acceptLanguage.get(language);
-	language = language.substring(0, 2);
-	//console.log(language);
-	return language;
-}
-
 var getStatic = function(req, res, type) {
-  var language = getRequestLanguage(req);
+  var language = langHelper.getRequestedLanguage(req);
   db.sql(sqlHelper.conv_i18n('SELECT TypeNumber, Title[_lang] from Static WHERE TypeID = ' + type + ';', language))
   .execute()
   .then(function (results) {
